Add chapter link support
The MangaDex API wrapper I found doesn't request any of the extra data,
nor does it support passing the `includes[]` parameter when fetching a
chapter (despite including that param in its method docstring), so the
only way I could see to output a chapter's parent manga title was to
fetch the manga object too.

Sigh. But it does work, just inefficiently (two requests for something
that should be doable in one).